requirements.txt
requests~=2.32.3
asyncio~=3.4.3
python-telegram-bot~=21.6
aiogram~=3.13.1
aiohttp~=3.10.10